refactor: lotto 테스트 패키지 이동 & 중복 확인 메서드 분리
