+++ README.md
@@ -10,10 +10,10 @@
 
 - [x] indent(인덴트, 들여쓰기) depth를 3이 넘지 않도록 구현한다. 2까지만 허용한다.
 - [x] 3항 연산자를 쓰지 않는다.
-- [ ] 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다. 함수(또는 메서드)가 한 가지 일만 잘 하도록 구현한다.
+- [x] 함수(또는 메서드)의 길이가 15라인을 넘어가지 않도록 구현한다. 함수(또는 메서드)가 한 가지 일만 잘 하도록 구현한다.
 - [x] else 예약어를 쓰지 않는다.
 - [x] Enum을 적용하여 프로그램을 구현한다.
-- [ ] UI(System.out, System.in, Scanner) 로직은 제외하고 단위테스트를 작성한다.
+- [x] UI(System.out, System.in, Scanner) 로직은 제외하고 단위테스트를 작성한다.
 - [x] 제공된 Lotto 클래스를 사용하여 구현해야 한다.
 - [x] Lotto에 numbers 이외의 필드(인스턴스 변수)를 추가할 수 없다. 
 - [x] numbers의 접근 제어자인 private은 변경할 수 없다.
